backend/requirements.txt
# =============================================================================
# Freelancer Feedback Assistant - Backend Dependencies
# =============================================================================
# Python 3.11+ required
# =============================================================================

# -----------------------------------------------------------------------------
# Web Framework
# -----------------------------------------------------------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# Database
# -----------------------------------------------------------------------------
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0

# -----------------------------------------------------------------------------
# AI & Machine Learning
# -----------------------------------------------------------------------------
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
tiktoken==0.5.1

# -----------------------------------------------------------------------------
# Task Queue
# -----------------------------------------------------------------------------
celery==5.3.4
redis==5.0.1
flower==2.0.1

# -----------------------------------------------------------------------------
# Email & Notifications
# -----------------------------------------------------------------------------
sendgrid==6.11.0
python-slack-sdk==3.26.1

# -----------------------------------------------------------------------------
# HTTP Client
# -----------------------------------------------------------------------------
httpx==0.25.2
aiohttp==3.9.1

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0

# -----------------------------------------------------------------------------
# Monitoring & Logging
# -----------------------------------------------------------------------------
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
structlog==23.2.0

# -----------------------------------------------------------------------------
# File Handling
# -----------------------------------------------------------------------------
python-magic==0.4.27
pillow==10.1.0

# -----------------------------------------------------------------------------
# Validation & Serialization
# -----------------------------------------------------------------------------
pydantic[email]==2.5.0
orjson==3.9.10

# -----------------------------------------------------------------------------
# CORS & Middleware
# -----------------------------------------------------------------------------
python-cors==1.0.0

# -----------------------------------------------------------------------------
# Rate Limiting
# -----------------------------------------------------------------------------
slowapi==0.1.9

# -----------------------------------------------------------------------------
# Performance
# -----------------------------------------------------------------------------
uvloop==0.19.0
httptools==0.6.1
